huggingface-agents-course/smolagents/pyproject.toml
[project]
name = "smollagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argparse>=1.4.0",
    "duckduckgo-search>=7.5.5",
    "geopandas>=1.0.1",
    "gradio-client>=1.8.0",
    "helium>=5.1.1",
    "huggingface-hub>=0.29.3",
    "kaleido==0.1.0.post1",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "numpy>=2.2.4",
    "openinference-instrumentation-smolagents>=0.1.8",
    "opentelemetry-exporter-otlp>=1.31.1",
    "opentelemetry-sdk>=1.31.1",
    "plotly>=6.1.1",
    "python-dotenv>=1.1.0",
    "rank-bm25>=0.2.2",
    "selenium>=4.33.0",
    "shapely>=2.1.1",
    "smolagents[litellm,mcp]>=1.12.0",
]
